Allow to override message settings
Ignore-this: 46079b8ee62ebfc63683208d81dd5468

In order to be able to control what is displayed when using the OASIS library,
we introduce a context parameter that hold the settings. This allow to control
when messages can be displayed and when they cannot.

We fallback to default parameter, when using it in Base. This module should
remain simple and can use global default settings.

darcs-hash:20100717234124-a43a9-22441513bf41434b6db24e1ba02a2efdbef124b0