Add {Interface,Implementation}Patterns to locate source files.
This change allows to override the built-in source files search. It uses
user-provided InterfacePatterns and ImplementationPatterns fields and a small
DSL to define how to find a file matching a given module. The fields can
be defined at the top level or in the Library section.

The DSL is based on Genlex and Buffer.add_substitute. It has two
functions `capitalize_file` and `uncapitalize_file` and one variable
`module`.

The new feature is "source_patterns":

    AlphaFeatures: source_patterns

    Library library
      [...]
      Modules: FooBar
      InterfacePatterns+: ${module}.eliomi
      ImplementationPatterns+:  ${module}.eliom

It is particularly useful to get rid of the warnings like:
W: Cannot find source file matching module [...]

Thanks to Petter Urkedal for this contribution.