Handle 'directory' field of findlib when installing.
Findlib allows to specify a "directory" in a META file. This can be used
to store some files of the library in sub-directories. This change
allows to set and use this "directory" field.

The InternalInstall plugin now handles installing files in sub-directories.
However there is no flag to specify this when using `ocamlfind install`, so we
use a standard file installation, with a best-guess computation of the target
directory.

The META plugin can now set the "directory" field according to what has
been set in the `_oasis` file. Note that, for nested package the install
directory is relative to the install directory of the parent package.

The new feature is "findlib_directory":

    BetaFeatures: findlib_directory
    Plugins: META(0.4)

    Library library
      [...]
      FindlibDirectory: lib

Thanks to Drup for this contribution. (Closes: #1305)