Replace OASISRecDescParser by OASISAst_parser using ocamlyacc.
This change has been tested against all the available OASIS files in
test/data/ and examples/ and it gives almost the same results. The only
exceptions are the AST of expressions which are now built using a
different priority in grammar rules.

The new LALR parser is faster than the old recurse descent (~40%
faster), but the speed-up is not really relevant overall.

Related refactorings:
* Replace TSLibrary, TSObject... by TSSection(section kind, ...)
* Add a benchmark
* Split tests in between TestOASISAst and TestOASISParse
* Introduce OASISMutex to create a critical section for the parser
* Extraneous spaces in the _oasis file are now fatal.

(Closes: #298)