Fix OPAM segfault in AppVeyor.
See https://ci.appveyor.com/project/Chris00/oasis-9nf6b/build/1.0.52#L2190

Assuming this is the same segfault I encountered several days ago, it
occurs when using OPAM's built-in heuristic solver. It is possible to
avoid that by installing aspcud.

This commit uses scripts from the binaries project [1] to install OCaml
and OPAM binaries in Cygwin. The OPAM script installs aspcud.

This should also speed up the build, as it is no longer necessary to
recompile OPAM when not cached.

[1]: https://github.com/aantron/binaries