Use OCaml naming policy for executables and prevent BuildDeps leaks.
The OCaml naming scheme prevents the use of "-" in module name, hence we
get warnings whenever compiling "oasis-dist.ml". This change renames the
files that don't apply OCaml naming policy to solve these errors. We
also take a step further by using what should be the OCaml module name
as the name of the file (capitalization).

This change also isolates executable in their own directories to prevent
BuildDeps leaks. The tags used for build depenidencies are applied
directory wide and whenever two executables are in the same directory,
they get the same set of package tags. This can be a problem when the
executables don't depend on the same packages -- one of which can be
optional.