Create OASIFileSystem and use it to OASISContext.t.
The new module OASISFileSystem is a filesystem abstraction layer that allows to
better represents (through types) where the files are.

The very first use of it is to add in OASISContext.t as srcfs. The
filesystem object is itself limited to only use
OASISContext.source_filename, which allows to use types to enforce
location of files.

The various default_filename across the code are now typed according to
their location (e.g. "setup.log" and "setup.data" are
OASISContext.source_filename). The only way to actually open, read and
write the corresponding file is to apply methods provided in OASISFileSystem
object.

This change enforce the use of the context, so we have to propagate it
everywhere, rather than relying on BaseContext.default. This has the bad
side effect to require non-backward compatible changes, hence a big
addition to BaseCompat to offset the change.

This change also delay enough the evaluation of the location of
"setup.log" and "setup.data", which solves the bug 1473.

(Closes: #1473)