Remove 'lazy'-ness from BaseEnvLight.default_filename.
This change decouple BaseEnvLight.default_filename and
BaseEnv.default_filename in term of code and introduce an extra test to
check that both functions return the same value.

After analysing the code, the value is only used, where we can assume
the "chdir", side effect of '-C' for OASIS, has already been done, for
example in myocamlbuild.ml.

The use of "lazy" for default_filename has been introduced to allow
using '-C dir' and have 'setup.data' located in 'dir'.