Add a FindlibExtraFiles field to library and object.
This change allows to use a new `_oasis` field 'FindlibExtraFiles' to
install extra files along the matching library/object in findlib. This
requires the feature 'findlib_extra_files'.

Examples:
[...]
BetaFeatures: findlib_extra_files

Library library
  [...]
  FindlibName: foobar
  FindlibExtraFiles: myfile.txt

In this case `myfile.txt` will be installed in the same directory as the
'foobar' package.

Thanks to Evgenii Lepikhin for the PR. (Closes: #802)