Try to register only when a glob pattern matches at least one file (#102)
* Try to register only when a glob pattern matches at least one file

BaseBuilt.register expects a list of non-empty lists of filenames as its
fourth argument. When it detects an empty list of filenames, it
generates a warning like:

    W: Cannot find an existing alternative files among:

OCamldoc does not generate the style.css file when option -css-style is
given. Then, in OCamlbuildDocPlugin.doc_build, OASISFileUtil.glob for
*.css returns an empty list, which is, in turn, fed into
BaseBuilt.register, resulting in a meaningless warning.